Move message fetching logic to a separate library module
- Extract `getMessages` function from API route to `lib/messages.ts`
- Remove redundant Message interface definition
- Simplify imports in page and API route components
- Improve code organization by centralizing message-related utilities
