+++ lib/index.ts
@@ -1,0 +1,2 @@
+// Re-export from messages.ts
+export * from './messages';
